\name{GenerateFactorData}
\alias{GenerateFactorData}
\title{Simulates Data from a Factor Model.}
\usage{
GenerateFactorData(n = 1000, nF = 2, nX = 5, nM = 5, nD = 5,
  sigma2.var = c(0.5, 4), normalize = diag(1, nF),
  filename = "simulation", corrfactors = T)
}
\arguments{
  \item{n}{the number of observations}

  \item{nF}{the number of unobserved factors}

  \item{nX}{the number of unobserved covariates}

  \item{nM}{the number of continuous measures}

  \item{nD}{the number of binary measures}

  \item{sigma2.var}{the bounds on the uniform distribution
  sigma2 is drawn from}

  \item{normalize}{a matrix of normalizations for the
  system (on first nF equations)}

  \item{filename}{the stub used for outputting data and
  variable names}

  \item{corrfactors}{A T/F for if factors should be
  generated with correlation}
}
\value{
numeric vector giving number of characters in each element
of the character vector.  Missing strings have missing
length.
}
\description{
\code{GenerateFactorData} creates a system of equations
generated by observables and a latent factor system.
Factors can be correlated and outcome equations can be
binary or continuous.  As a side effect a data set and list
of variables are saved to your working directory!
}
\examples{
GenerateFactorData(n=100)
GenerateFactorData(n=100, nF=2, nX=3, nM=5, filename="mySim", corrfactors=F)
}

